Re-arranged unit tests. Issue #265
