+++ Tests/DICOM.Tests.csproj
@@ -94,6 +94,7 @@
     <Compile Include="Bugs\GH223.cs" />
     <Compile Include="Bugs\GH220.cs" />
     <Compile Include="Bugs\GH258.cs" />
+    <Compile Include="Bugs\GH265.cs" />
     <Compile Include="DicomDatasetTest.cs" />
     <Compile Include="DicomDatasetWalkerTest.cs" />
     <Compile Include="DicomDateRangeTest.cs" />
